Add files via upload
Since ancient times, people have been interested in the motion of planets. For example the new seaworthy and overland routes were the motion and relative positions of the moon , sun, constellation affect the earth. With development of astronomy there were been physical study about difficult interaction of the massive objects which are located at huge distances from each other and moving on orbits - planets of solar system. Solutions of these studies led to fast development of mathematical apparatus in this area thanks to works of such most famous scientists as Newton, Kepler, etc. In the presents time a task of the relative movement of the 3rd and more bodies have no analytical decision. However the well developed differential calculus [1] allows us to present  the numerical solution of this task including the computer modeling. “To solve the mass point motion problem we need differential equations for the motion. The way we derive these equations doesn’t matter”: [2, §11]. One way to describe the motion of bodies is to present it through the coherent graph [3]. I this work we present  the graph building through solving a differential equation of second order curves(14) , using theorem of the center of mass and transition from a consecutive number of relative coordinates to general system of coordinates [2, 4].